refactor: move action to general data
