+++ src/Domains/Connectors/SalesAssist/Enums/LeadCustomFieldEnum.php
@@ -8,6 +8,8 @@ enum LeadCustomFieldEnum: string
 {
     case VEHICLE_OF_INTEREST = 'vehicle_of_interest';
     case TRADE_IN = 'vehicle_trade_id';
+    case TRADE_IN_DATA = 'tradein_data';
+    case TRADE_IN_IMPORTED = 'tradein_imported';
     case DRIVERS_LICENSE = 'get_docs_drivers_license';
     case DRIVERS_LICENSE_IMAGE = 'driver_license_images';
     case CREDIT_APP = 'credit_app';
